requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain-ollama==0.1.2
langchain-core==0.1.0
pytesseract==0.3.10
pillow==10.0.0
